requirements.txt
PyExecJS==1.5.1
requests==2.26.0
loguru==0.6.0
pandas==1.3.5
Scrapy==2.6.1





